Create Wheel class to calculate wheel circumference as well as diameter.
It is now worth promoting Wheel from a mere struct to a class, as we now have
more wheel-like behaviour to implement.

This change is painless and easy because all wheel-like behaviour that was
in Gear was already isolated – we just had to remove the definition of the
Wheel struct from Gear.