Managing dependencies: An example of isolated external messages
Gear sends messages to self, as well as to Wheel i.e. wheel.diameter in

This embedded, external dependency is especially bad if #gear_inches is a
complex method, as things are more likely to break whenever a change is made.
Much better, then, to isolate the wheel.diameter in a method of its own.
The result?

1. #gear_inches only sends messages to self

2. Also has a DRY-ing effect