Composition: specify bike defaults in arrays used by PartsFactory
PartsFactory does need to have knowledge of structure of config arrays for
this to work (wouldn't using config hashes be better? Anyway.)

However, knowledge of what the different bike defaults are has been restored
in the form of the config arrays, and the responsibility of putting them
together is now placed on PartsFactory.

Previously, when we had MountainBikeParts, RoadBikeParts etc, the knowledge
of defaults and how to put them together were spread across different classes.